adding a bunch of projectile bags

+++ BossFightGame/Assets/Scripts/ScriptableObject/SOPlayerClass.cs
@@ -5,7 +5,7 @@ public class SOPlayerClass : ScriptableObject
 {
     public Sprite previewSprite, HoverSprite, PrimaryAbilityImage, SecondaryAbilityImage, SpecialAbilityImage, MobilityAbilityImage;
 
-    public float BasePrimaryDamage, BaseSecondaryDamage, BaseSpecialDamage;
+    public int BasePrimaryDamage, BaseSecondaryDamage, BaseSpecialDamage;
 
     public string ClassName, PrimaryAbilityName, SecondaryAbilityName, SpecialAbilityName, MobilityAbilityName;
